+++ CHANGELOG.md
@@ -11,6 +11,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## v5.35.3
 
+### Fixed
+
 - Fix `thecodingmachine/safe` version constraint
 
 ### Fixed