+++ CHANGELOG.md
@@ -14,6 +14,26 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Added query caching
 - Added Apollo Automatic persisted queries support
 
+## v5.35.3
+
+### Fixed
+
+- Fix `thecodingmachine/safe` version constraint
+
+### Fixed
+
+## v5.35.2
+
+### Fixed
+
+- Ensure compatibility with `thecodingmachine/safe:2` https://github.com/nuwave/lighthouse/pull/2028
+
+## v5.35.1
+
+### Fixed
+
+- Fix partitioning of arguments where the model has a non-relation method returning primitive https://github.com/nuwave/lighthouse/pull/2026
+
 ## v5.35.0
 
 ### Added